+++ src/main/java/net/billstark001/worldmirror/xaero/WorldMirrorMixinPlugin.java
@@ -1,0 +1,42 @@
+package net.billstark001.worldmirror.xaero;
+
+import net.fabricmc.loader.api.FabricLoader;
+import org.objectweb.asm.tree.ClassNode;
+import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
+import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
+
+import java.util.List;
+import java.util.Set;
+
+public class WorldMirrorMixinPlugin implements IMixinConfigPlugin {
+    private static final String XAERO_GUI_MAP_MIXIN = "net.billstark001.worldmirror.xaero.mixin.XaeroGuiMapMixin";
+
+    @Override
+    public void onLoad(String mixinPackage) {}
+
+    @Override
+    public String getRefMapperConfig() {
+        return null;
+    }
+
+    @Override
+    public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+        if (!XAERO_GUI_MAP_MIXIN.equals(mixinClassName)) return true;
+        return FabricLoader.getInstance().isModLoaded("xaeroworldmap");
+    }
+
+    @Override
+    public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {}
+
+    @Override
+    public List<String> getMixins() {
+        return null;
+    }
+
+    @Override
+    public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {}
+
+    @Override
+    public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {}
+
+}